feat: multi-tenant Streamable-HTTP transport
Directories like Smithery can only introspect a server they can reach over
HTTP. A local stdio bundle is opaque to them, so it lists with no capabilities
however good the manifest is.

The obvious shortcut — wrap the existing binary in mcp-proxy and expose it —
is unsafe, and that is the whole point of this change. The stdio server reads
ONE key from the environment into ONE long-lived client shared by all 17 call
sites. Hosted that way it would serve every anonymous caller on the operator's
key, at the operator's tier: a silent, permanent giveaway of paid data. There
is no configuration of that binary which is safe to expose.

So the key becomes a parameter and the tools move behind a factory:

  server.ts  createServer(apiKey) — every tool, transport-agnostic. The client
             and all tool closures are created inside it, so two servers built
             from two keys share no mutable state.
  index.ts   stdio. One key from the environment. Single-tenant, which is
             correct for one user on one machine.
  http.ts    one createServer() PER REQUEST, bound to that caller's key, with
             the transport stateless (sessionIdGenerator: undefined) so there
             is no session map that could hand one caller another's server.

http.ts deliberately has no fallback to process.env. A missing caller key
yields an empty-key server — which every tool already handles by explaining
how to get one — rather than quietly borrowing the operator's credentials.
That fallback IS the bug this file exists to avoid.

test/http-isolation.mjs pins that property: 12 tools listable unauthenticated
(what an indexer needs), no env-key fallback, caller keys honoured via both
Authorization and X-API-Key, and no session id issued. Mutation-checked by
re-adding the env fallback, which fails the isolation assertion.

The test also refuses to run if something else already holds its port. A
leftover server from an earlier run answered for several minutes and made a
clean build look broken; the guard makes that misdiagnosis impossible rather
than relying on remembering to clean up.

stdio is unchanged and still passes test/protocol.mjs.
