+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/http.ts'],
   format: ['esm'],
   target: 'es2022',
   clean: true,
